requirements.txt
groq==0.7.0
sentence-transformers==3.0.1
scikit-learn==1.5.1
numpy==1.26.4
playwright==1.44.0
feedparser==6.0.11
requests==2.32.3
apscheduler==3.10.4
streamlit==1.36.0
pandas==2.2.2
chromadb==0.5.3
python-dotenv==1.0.1
pyTelegramBotAPI==4.21.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.2.2
httpx==0.27.0